% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{beta_muvat2shape1shape2}
\alias{beta_muvat2shape1shape2}
\title{Reparameterise Beta distributions}
\usage{
beta_muvat2shape1shape2(mu, sigma2)
}
\arguments{
\item{mu}{Mean, should be less than 1}

\item{sigma2}{variance, should be less than 1/12}
}
\value{
A named list containing 'shape1' and 'shape2'
}
\description{
Reparameterise Beta distributions
}
\author{
Sangeeta Bhatia
}
