% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{beta_shape1shape22muvar}
\alias{beta_shape1shape22muvar}
\title{Reparameterise Beta distributions}
\usage{
beta_shape1shape22muvar(shape1, shape2)
}
\arguments{
\item{shape1}{positive}

\item{shape2}{positive}
}
\value{
A named list containing 'mu' and 'sigma2'
}
\description{
Reparameterise Beta distributions
}
\author{
Sangeeta Bhatia
}
