% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dmbeta}
\alias{dmbeta}
\title{Modified beta distribution}
\usage{
dmbeta(x, shape1, shape2, xmax, ncp = 0, log = FALSE)
}
\arguments{
\item{x}{numeric value for which density is required}

\item{shape1}{non-negative parameter of the Beta distribution}

\item{shape2}{non-negative parameter of the Beta distribution}

\item{xmax}{Maximum possible value of x}

\item{ncp}{non-centrality parameter.}

\item{log}{logical; if TRUE, probabilities p are given as log(p)}
}
\value{
modified density
}
\description{
Modified beta probability density function
}
\details{
This function returns the probability density of the random
variable x/xmax where x/xmax is distributed according to a beta
distribution.
}
\author{
Sangeeta Bhatia
}
