% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_fast.R
\name{fast_probability_isolation_offset}
\alias{fast_probability_isolation_offset}
\title{Probability of serial interval when primary case has been
isolated and in the presence of asymptomatic infectiousness}
\usage{
fast_probability_isolation_offset(poffset, nu, offset, inf_params)
}
\arguments{
\item{poffset}{Probability calculation with offset only}

\item{nu}{onset to isolation of primary case}

\item{offset}{days of asymptomatic infectiousness}

\item{inf_params}{named list of arguments for
infectious period distribution.}
}
\value{
numeric probability of observing the given serial interval
with the given parameters of infectious period and incubation
period distribution
}
\description{
Probability of serial interval when primary case has been
isolated and in the presence of asymptomatic infectiousness
}
\author{
Sangeeta Bhatia
}
