% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pmbeta}
\alias{pmbeta}
\title{Modififed beta cumulative probability distribution function}
\usage{
pmbeta(x, shape1, shape2, xmax, ncp = 0, log = FALSE)
}
\arguments{
\item{x}{numeric value for which density is required}

\item{shape1}{non-negative parameter of the Beta distribution}

\item{shape2}{non-negative parameter of the Beta distribution}

\item{xmax}{Maximum possible value of x}

\item{ncp}{non-centrality parameter.}

\item{log}{logical; if TRUE, probabilities p are given as log(p)}
}
\value{
modified distribution
}
\description{
Modififed beta cumulative probability distribution function
}
\author{
Sangeeta Bhatia
}
