% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_si}
\alias{simulate_si}
\title{Simulate serial interval}
\usage{
simulate_si(
  mean_ip,
  sd_ip,
  mean_inf,
  sd_inf,
  mean_iso = NULL,
  sd_iso = NULL,
  offset = 0,
  nsim = 50
)
}
\arguments{
\item{mean_ip}{Mean Incubation Period}

\item{sd_ip}{Standard deviation of the incubation period}

\item{mean_inf}{Infectious Period}

\item{sd_inf}{Standard deviation of the infectious period}

\item{mean_iso}{Mean delay from symptom onset to case isolation.
If this is null, then isolation is not simulated.}

\item{sd_iso}{Standard deviation of the delay from symptom onset
to case isolation}

\item{offset}{Numeric - days before symptom onset that
infectiousness is assumed to begin. If \code{offset} is NULL, then
asymptomatic infectiousness is assumed to be absent.}

\item{nsim}{Number of transmission pairs to simulate}
}
\value{
a data frame with columns t_1, t_2, si
t_1 corresponds to draws from the infectious period,
t_2 corresponds to draws from the incubation period, and
si is the serial interval (t_1 + t_2).
If mean_iso and sd_iso are not NULL, then the output also contains
a column nu corresponding to draws from the distribution of onset
to isolation.
}
\description{
Simulate serial interval
}
\details{
This function simulates serial interval with or without case
isolation, and with or without pre-symptomatic infectiousness.
At the moment, infectious period, incubation period and onset to
isolation distributions are assumed to be gamma distributed.
}
\author{
Sangeeta Bhatia
}
