% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{total_log_likelihood}
\alias{total_log_likelihood}
\title{Total log likelihood}
\usage{
total_log_likelihood(
  tvec,
  nu_vec = NULL,
  offset_vec = NULL,
  inf_params,
  ip_params
)
}
\arguments{
\item{tvec}{Vector of observed serial intervals}

\item{nu_vec}{Vector of delays from onset to isolation}

\item{offset_vec}{Vector specifying the pre-symptomatic infectiousness}

\item{inf_params}{list with components rate and shape for
infectious period distribution}

\item{ip_params}{list with components rate and shape for
incubation period distribution}
}
\value{
total log-likelihood
}
\description{
Total log-likelhood
}
\details{

}
\author{
Sangeeta Bhatia
}
