% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{probability_basic}
\alias{probability_basic}
\title{Probability of serial interval as a convolution of
infectiousness distribution of primary case and incubation period
of secondary case}
\usage{
probability_basic(t, inf_params, ip_params)
}
\arguments{
\item{t}{observed serial interval}

\item{inf_params}{list with components rate and shape for
infectious period distribution}

\item{ip_params}{list with components rate and shape for
incubation period distribution}
}
\value{
numeric. Probability of observing the given serial
interval given
the parameters of infectious period distributions
and incubation period distributions
}
\description{
Probability of serial interval as a convolution of
infectiousness distribution of primary case and incubation period
of secondary case
}
\author{
Sangeeta Bhatia
}
