% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{probability_isolation}
\alias{probability_isolation}
\title{Probability of serial interval}
\usage{
probability_isolation(t, nu, inf_params, ip_params)
}
\arguments{
\item{t}{observed serial interval}

\item{nu}{onset to isolation of primary case}

\item{inf_params}{list with components rate and shape for
infectious period distribution}

\item{ip_params}{list with components rate and shape for
incubation period distribution}
}
\value{
numeric. Probability of observing the given serial
interval given the delay from onset to isolation,
the parameters of infectious period distributions
and incubation period distributions.
}
\description{
Probability of serial interval
}
\author{
Sangeeta Bhatia
}
