% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{probability_offset}
\alias{probability_offset}
\title{Probability of serial interval
in the presence of asymptomatic infectiousness}
\usage{
probability_offset(t, offset, inf_params, ip_params)
}
\arguments{
\item{t}{observed serial interval}

\item{offset}{days of asymptomatic infectiousness}

\item{inf_params}{list with components rate and shape for
infectious period distribution}

\item{ip_params}{list with components rate and shape for
incubation period distribution}
}
\value{
numeric probability of observing the given serial interval
with the given parameters of infectious period and incubation
period distribution
}
\description{
Probability of serial interval
in the presence of asymptomatic infectiousness
}
\author{
Sangeeta Bhatia
}
